Make it compatible with freerdp master as of 21-03-2023.

+++ src/plugins/platforms/freerdp/qfreerdppeer.h
@@ -72,7 +72,7 @@ protected:
 	static BOOL xf_mouseEvent(rdpInput *input, UINT16 flags, UINT16 x, UINT16 y);
 	static BOOL xf_extendedMouseEvent(rdpInput *input, UINT16 flags, UINT16 x, UINT16 y);
 	static BOOL xf_input_synchronize_event(rdpInput *input, UINT32 flags);
-	static BOOL xf_input_keyboard_event(rdpInput *input, UINT16 flags, UINT16 code);
+	static BOOL xf_input_keyboard_event(rdpInput *input, UINT16 flags, UINT8 code);
 	static BOOL xf_input_unicode_keyboard_event(rdpInput *input, UINT16 flags, UINT16 code);
 	static BOOL xf_refresh_rect(rdpContext *context, BYTE count, const RECTANGLE_16* areas);
 	static BOOL xf_suppress_output(rdpContext* context, BYTE allow, const RECTANGLE_16 *area);
